requirements.txt
dash==2.11.0
pandas==2.3.2
numpy==2.2.6
plotly==5.21.0
requests==2.31.0
python-dotenv==1.0.1
scikit-learn==1.3.2
joblib==1.3.2
Flask==2.2.5
